requirements.txt
Django==2.1.2
twilio==6.20.0
Pillow==5.3.0
hashids==1.2.0
django-storages==1.7.1